+++ packages/signing/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vocdoni/signing - Changelog
 
+## 1.16.0
+
+- `sortJson` is now exported as `JsonSignature.sort`
+- `JsonSignature` and `BytesSignature` now have `signMessage`, `signTransaction`, `isValidMessage`, `isValidTransaction`, `recoverMessagePublicKey` and `recoverTransactionPublicKey`
+- Breaking:
+  - `sign`, `isValid` and `recoverPublicKey` no longer exist on `JsonSignature` or `BytesSignature`
+  - The signature of `isValidMessage`/`isValidTransaction` is now consistent with `recoverMessagePublicKey`/`recoverTransactionPublicKey`
+
+## 1.15.1
+
+- Exposing `sortJson`, `digestVocdoniSignedPayload` and `normalizeJsonToString`
+
 ## 1.15.0
 
 - Adding support for Vocdoni salted signatures